changed error handling and moved parse_info struct
