+++ backend/src/interface/mensa_parser.rs
@@ -5,14 +5,6 @@ use crate::{interface::mensa_parser::model::ParseCanteen, util::Date};
 use async_trait::async_trait;
 use thiserror::Error;
 
-#[derive(Debug)]
-pub struct ParseInfo {
-    pub base_url: String,
-    pub valid_canteens: Vec<String>,
-    pub client_timeout: std::time::Duration,
-    pub client_user_agent: String,
-}
-
 #[derive(Debug, Error)]
 pub enum ParseError {
     #[error("the node was not found: {0}")]
